+++ src/components/ActionCollection.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref } from 'vue'
 import actionStore from '@/actionStore'
-import AddActionDialog from './AddActionDialog.vue'
+import ActionDialog from './ActionDialog.vue'
 import ActionComponent from './ActionComponent.vue'
 
 const actions = ref(actionStore.getActions())
@@ -16,5 +16,5 @@ function actionsUpdatedEvent() {
     <q-list bordered separator v-for="action in actions" :key="action.Id">
         <ActionComponent @actions-updated="actionsUpdatedEvent" :action="action" />
     </q-list>
-    <AddActionDialog @actions-updated="actionsUpdatedEvent" />
+    <ActionDialog @actions-updated="actionsUpdatedEvent" />
 </template>
