+++ src/components/ActionComponent.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import AddActionDialog from './AddActionDialog.vue'
+import ActionDialog from './ActionDialog.vue'
 import RemoveActionButton from './RemoveActionButton.vue'
 import type { Action } from '@/action'
 
@@ -12,7 +12,7 @@ const emit = defineEmits(['actions-updated'])
         <q-item-section>
             <q-item-label>{{ action.Label }}</q-item-label>
         </q-item-section>
-        <AddActionDialog @actions-updated="emit('actions-updated')" :action />
+        <ActionDialog @actions-updated="emit('actions-updated')" :action />
         <RemoveActionButton
             @actions-updated="emit('actions-updated')"
             :action-id="props.action.Id"
